Fix possible client freeze on 1.9 and earlier.

+++ src/game/Spells/Spell.h
@@ -710,7 +710,7 @@ class Spell
                 struct
                 {
                     uint32 amount;
-                    uint32 critical;
+                    uint8 critical;
                 } heal;
 
                 struct
